# docs/04-api/openapi.yaml
openapi: 3.0.3
info:
  title: Platinum Casino API
  description: |
    REST API for the Platinum Casino online gaming platform.

    Authentication is handled via HTTP-only cookies containing a JWT token.
    After a successful login or registration, the server sets an `authToken` cookie
    that is automatically sent with subsequent requests.

    Rate limiting is applied globally to all `/api` routes, with stricter limits
    on authentication and bet-placement endpoints.
  version: 1.0.0
  contact:
    name: Platinum Casino Dev Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
...
security:
  - cookieAuth: []
paths:
...
  /api/users/me:
    get:
      tags: [Users]
      summary: Get current user data
      description: Returns the authenticated user's full profile including balance and activity timestamps.
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/users/profile:
    get:
      tags: [Users]
      summary: Get user profile
      description: Returns the authenticated user's profile with balance from the balance history table.
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      role:
                        $ref: '#/components/schemas/UserRole'
                      balance:
                        type: number
                        format: double
                      createdAt:
                        type: string
                        format: date-time
                      lastLogin:
                        type: string
                        format: date-time
                        nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Users]
      summary: Update user profile (change password)
      description: Allows the authenticated user to change their password by providing the current password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  description: The user's current password
                newPassword:
                  type: string
                  description: The desired new password
              required:
                - currentPassword
                - newPassword
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
        '400':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/users/balance:
    get:
      tags: [Users]
      summary: Get current balance
      description: Returns the authenticated user's latest balance.
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    format: double
                    example: 1250.50
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/users/balance/history:
    get:
      tags: [Users]
      summary: Get balance history
      description: Returns the authenticated user's full balance change history.
      responses:
        '200':
          description: List of balance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/users/transactions:
    get:
      tags: [Users]
      summary: Get user transactions
      description: Returns transactions for the authenticated user with optional filters.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of transactions to return
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
          description: Filter by transaction type
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter transactions on or after this date
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter transactions on or before this date
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---------------------------------------------------------------------------
  # Games
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authToken
      description: |
        JWT stored in an HTTP-only cookie. Automatically set by the server on
        successful login or registration. The token expires after 24 hours.
        The JWT payload contains `userId`, `username`, and `role`.
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          example:
            message: No token provided
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
  schemas:
...
    UserRole:
      type: string
      enum: [user, admin]
      default: user
    TransactionType:
      type: string
      enum: [deposit, withdrawal, game_win, game_loss, admin_adjustment, bonus, login_reward]
    TransactionStatus:
      type: string
      enum: [pending, completed, failed, voided, processing]
    GameType:
      type: string
      enum: [crash, plinko, wheel, roulette, blackjack]
    BalanceChangeType:
      type: string
      enum: [deposit, withdrawal, win, loss, admin_adjustment, login_reward]
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          example: An error occurred
      required:
        - message
...
    UserProfile:
      type: object
      description: User profile returned by /api/users/me
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: player1
        role:
          $ref: '#/components/schemas/UserRole'
        balance:
          type: number
          format: double
          example: 1250.50
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
...
    Transaction:
      type: object
      description: Financial transaction record
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          $ref: '#/components/schemas/TransactionType'
        gameType:
          $ref: '#/components/schemas/GameType'
        amount:
          type: string
          description: Decimal stored as string (precision 15, scale 2)
        balanceBefore:
          type: string
          description: Balance before transaction
        balanceAfter:
          type: string
          description: Balance after transaction
        status:
          $ref: '#/components/schemas/TransactionStatus'
        createdBy:
          type: integer
          nullable: true
          description: Admin user ID if manually created
        reference:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        gameSessionId:
          type: integer
          nullable: true
        metadata:
          type: object
          nullable: true
        notes:
          type: object
          nullable: true
        voidedBy:
          type: integer
          nullable: true
        voidedReason:
          type: string
          nullable: true
        voidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BalanceRecord:
      type: object
      description: Single balance-change entry
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: string
          description: Balance after the change
        previousBalance:
          type: string
          description: Balance before the change
        changeAmount:
          type: string
          description: Amount added or subtracted
        type:
          $ref: '#/components/schemas/BalanceChangeType'
        gameType:
          $ref: '#/components/schemas/GameType'
        relatedSessionId:
          type: integer
          nullable: true
        transactionId:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
        adminId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...